wallet: split wallet db implementations to specific files
